+++ examples/nuxt/server/lucia.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="lucia-auth" />
-declare namespace Lucia {
-	type Auth = import("./utils/auth.js").Auth;
-	type UserAttributes = {
-		username: string;
-	};
-}