+++ packages/lucia/src/auth/request.ts
@@ -32,12 +32,18 @@ export class AuthRequest<A extends Auth = any> {
 		this.auth = auth;
 		this.context = context;
 		try {
-			auth.validateRequestOrigin(context.request);
-			this.storedSessionId = auth.readSessionCookie(context.request);
+			if (auth.csrfProtectionEnabled) {
+				auth.validateRequestOrigin(context.request);
+			}
+			this.storedSessionId =
+				context.request.storedSessionCookie ??
+				auth.readSessionCookie(context.request.headers.cookie);
 		} catch (e) {
 			this.storedSessionId = null;
 		}
-		this.bearerToken = auth.readBearerToken(context.request);
+		this.bearerToken = auth.readBearerToken(
+			context.request.headers.authorization
+		);
 	}
 
 	private validatePromise: Promise<Session | null> | null = null;